requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
flasgger==0.9.5
Flask==2.0.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.3
MarkupSafe==2.0.1
mistune==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
